net/: Import Huawei kernel changes
* From Kobe2_10.1.0_Code_Opensource.tar.gz

+++ net/netfilter/xt_quota2.c
@@ -30,13 +30,13 @@
 
 /* Format of the ULOG packets passed through netlink */
 typedef struct ulog_packet_msg {
-	unsigned long mark;
-	long timestamp_sec;
-	long timestamp_usec;
+	unsigned int mark;
+	int timestamp_sec;
+	int timestamp_usec;
 	unsigned int hook;
 	char indev_name[IFNAMSIZ];
 	char outdev_name[IFNAMSIZ];
-	size_t data_len;
+	unsigned int data_len;
 	char prefix[ULOG_PREFIX_LEN];
 	unsigned char mac_len;
 	unsigned char mac[ULOG_MAC_LEN];
